requirements.txt
opencv-python
ultralytics
opencv-python
google-generativeai
httpx
modal
pyttsx3